Fail template rename job if push fails
This make the tempalte renamer job fail if the push action was not able to push. 

This will normally happen after the project was forked. When the template name was changed during or after the fork the template renamer can only do its work if it has writing rights. This check shall make a hint what is wrong.